add technical questions demo
solved the technical questions:

Given an array of integers, write a method to total the odd numbers.

Given an array of integers, write a method to sum the elements in the array, knowing that some of the elements may be very large integers.

Given a string, reverse it.

Given a string, remove any repeated letters.

FizzBuzz